Fix test broken in gh-6285

+++ theano/gof/tests/test_graph_opt_caching.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import, print_function, division
-import os
+import os.path
 import numpy as np
 import theano
 import theano.tensor as T
@@ -9,7 +9,8 @@ floatX = 'float32'
 
 def test_graph_opt_caching():
     opt_db_file = os.path.join(theano.config.compiledir, 'optimized_graphs.pkl')
-    os.remove(opt_db_file)
+    if os.path.exists(opt_db_file):
+        os.remove(opt_db_file)
 
     mode = theano.config.mode
     if mode in ["DEBUG_MODE", "DebugMode"]:
